fix: finalizeTmpImageIfNeeded를 공통 유틸로 빼고 경로 순회(../) 방어
