+++ src/app/api/admin/products/[id]/route.ts
@@ -3,11 +3,9 @@ import { createClient } from '@/app/lib/supabase/server';
 import { createAdminClient } from '@/app/lib/supabase/admin';
 import { getAdminUser } from '@/lib/admin-auth';
 import { getProductTableByCategory } from '@/lib/admin-product';
+import { finalizeTmpImageIfNeeded, getSafeStoragePathFromPublicUrl } from '@/lib/admin-product-image';
 import {
-  ADMIN_PRODUCT_LIVE_PREFIX,
-  ADMIN_PRODUCT_TMP_PREFIX,
   PRODUCT_IMAGE_BUCKET,
-  SUPABASE_URL,
 } from '@/constants';
 
 type UpdateBody = {
@@ -19,49 +17,6 @@ type UpdateBody = {
 
 const unauthorized = () =>
   NextResponse.json({ error: '관리자 권한이 필요합니다.' }, { status: 401 });
-
-const getStoragePathFromPublicUrl = (
-  imageUrl: string | null | undefined,
-  bucket: string,
-) => {
-  if (!imageUrl) return null;
-
-  const marker = `/storage/v1/object/public/${bucket}/`;
-  const markerIndex = imageUrl.indexOf(marker);
-  if (markerIndex === -1) return null;
-
-  const rawPath = imageUrl.slice(markerIndex + marker.length).split('?')[0];
-  if (!rawPath) return null;
-  return decodeURIComponent(rawPath);
-};
-
-const getPublicUrlFromPath = (path: string) => {
-  if (!SUPABASE_URL) return '';
-  return `${SUPABASE_URL}/storage/v1/object/public/${PRODUCT_IMAGE_BUCKET}/${path}`;
-};
-
-const finalizeTmpImageIfNeeded = async (imageUrl: string) => {
-  const imagePath = getStoragePathFromPublicUrl(imageUrl, PRODUCT_IMAGE_BUCKET);
-  if (!imagePath || !imagePath.startsWith(ADMIN_PRODUCT_TMP_PREFIX)) {
-    return imageUrl;
-  }
-
-  const extension = imagePath.split('.').pop() || 'png';
-  const finalPath = `${ADMIN_PRODUCT_LIVE_PREFIX}${Date.now()}-${Math.random()
-    .toString(36)
-    .slice(2, 9)}.${extension}`;
-
-  const adminClient = createAdminClient();
-  const { error } = await adminClient.storage
-    .from(PRODUCT_IMAGE_BUCKET)
-    .move(imagePath, finalPath);
-
-  if (error) {
-    throw new Error(`임시 이미지 확정 실패: ${error.message}`);
-  }
-
-  return getPublicUrlFromPath(finalPath);
-};
 
 export async function PATCH(
   req: Request,
@@ -138,11 +93,11 @@ export async function PATCH(
     );
   }
 
-  const beforeImagePath = getStoragePathFromPublicUrl(
+  const beforeImagePath = getSafeStoragePathFromPublicUrl(
     currentProduct?.image,
     PRODUCT_IMAGE_BUCKET,
   );
-  const afterImagePath = getStoragePathFromPublicUrl(
+  const afterImagePath = getSafeStoragePathFromPublicUrl(
     data?.image,
     PRODUCT_IMAGE_BUCKET,
   );
@@ -204,7 +159,7 @@ export async function DELETE(
     );
   }
 
-  const imagePath = getStoragePathFromPublicUrl(
+  const imagePath = getSafeStoragePathFromPublicUrl(
     target?.image,
     PRODUCT_IMAGE_BUCKET,
   );
